requirements.txt
numpy
pandas
scikit-learn
dlib-bin

git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0

pillow
segno

supabase
bcrypt

streamlit

librosa
resemblyzer
